refactor: 코드래빗 PR 8 추가 반영 — InvalidEnumException + SearchCriteria port.in 이동 + Wallet catch 정리 + 테스트 보강
