+++ src/main/java/com/trustamarket/orderservice/order/application/port/in/OrderSearchCriteria.java
@@ -1,11 +1,12 @@
-package com.trustamarket.orderservice.order.application.port.out.query;
+package com.trustamarket.orderservice.order.application.port.in;
 
-import com.trustamarket.orderservice.order.domain.exception.InvalidIdException;
+import com.trustamarket.orderservice.order.application.exception.InvalidSearchCriteriaException;
 import com.trustamarket.orderservice.order.domain.model.OrderStatus;
 
 import java.time.Instant;
 
 // SearchOrdersUseCase 검색 조건 (ADMIN)
+// 위치: port.in — UseCase의 입력. Repository.search()도 동일 record를 받음 (port.out → port.in 정방향 의존)
 // 모든 필드 nullable — null이면 해당 조건 무시
 // MVP scope: status + 기간(fromDate~toDate) + buyerName 부분일치
 // TODO: 추후 풍부한 검색(키워드 full-text + 다중 필터) 도입 시 확장
@@ -18,7 +19,7 @@ public record OrderSearchCriteria(
     public OrderSearchCriteria {
         // 기간 역전 차단
         if (fromDate != null && toDate != null && fromDate.isAfter(toDate)) {
-            throw new InvalidIdException("dateRange");   // TODO: 시각 범위 전용 예외 신설 검토
+            throw new InvalidSearchCriteriaException("dateRange", "fromDate must be <= toDate");
         }
         // buyerName blank → null 정규화 (조건 무시 의도와 일관)
         if (buyerName != null && buyerName.isBlank()) {
